libintelprov: info.lifetime.token_size is expressed in 32bits word
BZ: 130884

info.lifetime.token_size is expressed in 32bits word not in 8bits
word. read_token should compute the token size accordingly.

Reviewed-on: http://android.intel.com:8080/125592
Tested-by: Lumbierres, PhilippeX <philippex.lumbierres@intel.com>
Tested-by: cactus <cactus@intel.com>

+++ tee_connector.c
@@ -274,7 +274,7 @@ int read_token(int argc, char **argv)
 		return EXIT_FAILURE;
 	}
 
-	buf = (uint8_t *)malloc(info.lifetime.token_size * sizeof(uint8_t));
+	buf = (uint8_t *)malloc(info.lifetime.token_size * sizeof(uint32_t));
 	if (buf == NULL)
 	{
 		raise_error("Failed to alloc the buffer to retrieve the token data, error=%s", strerror(ENOMEM));
